Add a KeysInterface to bridge between user wallet and others lightning componenents, responsible to get seed from the former and derive appropriate keyring materials
Move ChannelKeys into keysinterface for generate a set of it from
master_seed
